Fix bugs

+++ src/main/java/zju/cst/aces/ant/Main.java
@@ -15,13 +15,18 @@ public class Main {
     static Path compiledPath = Paths.get("/Users/chenyi/Downloads/sf110-chatgpt/SF110-20130704");
     static Path dependencyPath = Paths.get("/Users/chenyi/Desktop/ChatUniTest/chatunitest-standalone/src/main/resources/dependency");
     static String apiKey = "xxx";
+    static String url = "http://127.0.0.1/v1/chat/completions";
+    static String model = "code-llama";
 
     public static void main(String[] args) {
         // 检查是否传递了参数
-        autoEnv4SF();
-
         if (args.length == 0) {
             System.out.println("Usage: java -jar my-project.jar [envFilePath] [arg1] [arg2]...");
+            return;
+        }
+
+        if (args[0].equals("createEnv")) {
+            autoEnv4SF(args);
             return;
         }
 
@@ -31,8 +36,24 @@ public class Main {
         new CommandHandler(envFilePath).handle(command);
     }
 
-    static void autoEnv4SF() {
-        for (File file : srcPath.toFile().listFiles()) {
+    static void autoEnv4SF(String[] args) {
+        if (args.length != 7) {
+            System.out.println("Usage: java -jar chatunitest-standalone.jar createEnv [srcPath] [compiledPath] [dependencyPath] [apiKey] [url] [model]");
+            return;
+        }
+        srcPath = Paths.get(args[1]);
+        compiledPath = Paths.get(args[2]);
+        dependencyPath = Paths.get(args[3]);
+        apiKey = args[4];
+        url = args[5];
+        model = args[6];
+
+        if (!srcPath.toFile().exists() || !compiledPath.toFile().exists() || !dependencyPath.toFile().exists()) {
+            System.out.println("srcPath, compiledPath or dependencyPath not exists");
+            return;
+        }
+
+        for (File file : Objects.requireNonNull(srcPath.toFile().listFiles())) {
             if (file.isDirectory()) {
                 if (!file.getName().contains("_")) {
                     continue;
@@ -46,11 +67,10 @@ public class Main {
                 try {
                     Path absProjectPath = file.toPath().toAbsolutePath();
                     content.append("apiKeys=").append(apiKey).append("\n");
-                    content.append("url=\n");
+                    content.append("url=").append(url).append("\n");
                     content.append("model=\n");
                     content.append("enableMultithreading=").append(true).append("\n");
                     content.append("parentEnvPath=\n");
-                    content.append("classPaths=").append(artifactId).append("\n");
                     content.append("baseDir=").append(absProjectPath).append("\n");
                     content.append("packaging=\n");
                     content.append("groupId=").append("com.example").append("\n");
@@ -75,11 +95,14 @@ public class Main {
 
     static void copyCompiledFiles(Path srcDirPath, String artifactId) {
         // copy compiled files
-        Path compiledProjectPath = compiledPath.resolve(artifactId);
+        Path compiledProjectPath = compiledPath.resolve(srcDirPath.toFile().getName());
         try {
             File srcDir = compiledProjectPath.toFile();
             File destDir = srcDirPath.resolve("target").toFile();
-            for (File file : srcDir.listFiles()) {
+            if (!destDir.exists() && !destDir.mkdirs()) {
+                throw new RuntimeException("Failed to create directory: " + destDir);
+            }
+            for (File file : Objects.requireNonNull(srcDir.listFiles())) {
                 if (file.isDirectory() && (file.getName().equals("target") || file.getName().equals("build"))) {
                     FileUtils.copyDirectory(file, destDir);
                 } else if (file.isFile() && file.getName().equals(artifactId + ".jar")) {
